+++ chaosprometheus/probes.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
+import statistics
+from datetime import datetime
 from typing import Any, Dict
 
 try:
@@ -6,14 +8,13 @@ try:
     import maya
     import requests
     from logzero import logger
-except ImportError as x:
-    print(x)
+except ImportError:
     raise
 
-from chaoslib.exceptions import FailedActivity
+from chaoslib.exceptions import ActivityFailed, FailedActivity
 from chaoslib.types import Configuration, Secrets
 
-__all__ = ["query", "query_interval"]
+__all__ = ["query", "query_interval", "compute_mean", "nodes_cpu_usage_mean"]
 
 
 def query(
@@ -109,3 +110,83 @@ def query_interval(
         )
 
     return r.json()
+
+
+def compute_mean(
+    query: str,
+    window: str = "1d",
+    mean_type: str = "arithmetic",
+    configuration: Configuration = None,
+    secrets: Secrets = None,
+) -> float:
+    """
+    Compute the mean of all returned datapoints of the range vector matching
+    the given query. The query must return a range vector.
+
+    The default computes an arithmetic mean. You can switch to geometric
+    or harmonic mean by passing `mean_type="geometric"` or
+    `mean_type="harmonic"`.
+    """
+    prom_url = (configuration or {}).get(
+        "prometheus_base_url", "http://localhost:9090"
+    )
+
+    url = f"{prom_url}/api/v1/query_range"
+
+    start_period = dateparser.parse(
+        window, settings={"RETURN_AS_TIMEZONE_AWARE": True}
+    )
+    today = datetime.today()
+    end_period = today.timestamp()
+    step = (end_period - start_period) / 60 / 5
+
+    r = requests.get(
+        url,
+        params={
+            "query": query,
+            "start": start_period,
+            "end": end_period,
+            "step": step,
+        },
+    )
+
+    if r.status_code > 399:
+        logger.debug(
+            f"Failed to perform Prometheus query '{query}': {r.json()}"
+        )
+        raise ActivityFailed("Failed to run Prometheus query")
+
+    response = r.json()
+    if response["status"] != "success":
+        logger.debug(f"Prometheus query '{query}' not successful: {response}")
+        raise ActivityFailed("Prometheus query was not successful")
+
+    data = list(
+        map(lambda i: float(i[1]), response["data"]["result"][0]["values"])
+    )
+
+    if mean_type == "geometric":
+        return statistics.geometric_mean(data)
+
+    elif mean_type == "geometric":
+        return statistics.harmonic_mean(data)
+
+    return statistics.fmean(data)
+
+
+def nodes_cpu_usage_mean(
+    window: str = "1d",
+    configuration: Configuration = None,
+    secrets: Secrets = None,
+) -> float:
+    """
+    Computes a mean of all nodes activities per minute over the given
+    `window`. We use the `node_cpu_seconds_total` metric to perform this
+    query.
+    """
+    return compute_mean(
+        '100 * (1 - avg(rate(node_cpu_seconds_total{mode="idle"}[1m])))',
+        window=window,
+        configuration=configuration,
+        secrets=secrets,
+    )