+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 [Unreleased]: https://github.com/chaostoolkit/chaostoolkit-prometheus/compare/0.4.0...HEAD
 
+### Added
+
+- Two new probes to compute the mean of a range vector
+
 ## [0.4.0][] - 2022-02-17
 
 [0.4.0]: https://github.com/chaostoolkit/chaostoolkit-prometheus/compare/0.3.0...0.4.0